move zingo-testvectors out
